feat: add dereferenceAliases to DumperOptions
Setting this property to true will force emitter to dereference aliases:
emit Node referenced by the Alias instead of Alias itself.
The Anchors are not emitted also.

Default value is true to preserve current behavior aligned with specs.

If serializer "detects" recursive structure it fails with
SerializationException.

Implementation is on the Serialize level meaning that Nodes are
always constructed with Anchors and Aliases.

There is no test with Merge Nodes.

see #1086