feat: merge nodes on Compose
Implement merge expansion while Composing nodes.

This does not really clone nodes which need to be inserted on MERGE.
It inserts same node whenever it is referenced. This should not be a
problem and I guess it works the same now when merge processed during
Construction.
If you dump loaded node tree without
DumperOptions.setDereferenceAliases(true) it may be unreadable, but
kind of still valid.

see #1096